+++ package.json
@@ -14,6 +14,7 @@
   "homepage": "https://fs-oct24-best-practice.github.io/best-practice-project/",
   "dependencies": {
     "clsx": "^2.1.1",
+    "express": "^4.21.2",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-router-dom": "^7.1.1"
