+++ src/pages/TabletsPage/TabletsPage.tsx
@@ -1,3 +1,7 @@
-export const TabletsPage = () => {
-	return <h1>Tablets Page</h1>;
+import React from 'react';
+import {Catalog} from '../../components/catalog';
+import {getTablets} from '../../components/api/apiE';
+
+export const TabletsPage: React.FC = () => {
+	return <Catalog fetchProducts={getTablets} title='Tablets' />;
 };
