+++ src/types/Product.ts
@@ -1,4 +1,5 @@
-export interface ProductDescription {
+export interface Product {
+	priceRegular: number;
 	id: number;
 	category: string;
 	itemId: string;
