+++ tsconfig.json
@@ -4,7 +4,6 @@
     "paths": {
       "@best-practice/images/*": ["public/mockup_img/*"],
       "@best-practice/pages/*": ["src/pages/*"],
-      "@best-practice/api/*": ["public/api/*"]
     }
   },
   "files": [],
