+++ vite.config.ts
@@ -13,7 +13,6 @@ export default defineConfig({
 		alias: {
 			'@best-practice/images': path.resolve(__dirname, 'public/mockup_img'),
 			'@best-practice/pages': path.resolve(__dirname, 'src/pages'),
-			'@best-practice/api': path.resolve(__dirname, 'public/api'),
 		},
 	},
 });
